+++ Engine/src/Engine.h
@@ -4,10 +4,10 @@
 // Do not use in Engine
 
 #include "Engine/Application.h"
-#include "Engine/Log.h"
-#include "Engine/ResourceManager.h"
-#include "Engine/Scene.h"
-#include "Engine/Serializer.h"
+#include "Engine/Utilities/Log.h"
+#include "Engine/ResourceManagement/ResourceManager.h"
+#include "Engine/SceneBuilder/Scene.h"
+#include "Engine/ResourceManagement/Serializer.h"
 
 // ----------- Entry Point -----------
 #include "Engine/EntryPoint.h"
