+++ Engine/src/Engine/ResourceManagement/ShaderGenerator.cpp
@@ -1,5 +1,5 @@
-#include "ShaderGenerator.h"
-#include "Log.h"
+#include <Engine/ResourceManagement/ShaderGenerator.h>
+#include "Engine/Utilities/Log.h"
 #include "glad/glad.h"
 
 namespace Engine
