+++ Engine/src/Engine/Scripts/ScriptableBehavior.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "Engine/Core.h"
-#include "Engine/Entity.h"
+#include "Engine/SceneBuilder/Entity.h"
 
 namespace Engine
 {
