+++ Engine/src/Engine/Utilities/Log.h
@@ -2,7 +2,7 @@
 
 #include <memory>
 
-#include "Core.h"
+#include "Engine/Core.h"
 #include "spdlog/spdlog.h"
 
 namespace Engine
